requirements-dev.txt
astroid==2.3.2
Django==2.2.7
django-cors-headers==3.2.0
django-debug-toolbar==2.1
django-otp==0.7.3
django-request-logging==0.7.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
PyJWT==1.7.1
pylint==2.4.3
python-decouple==3.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
